refactor feature annotations to package
