--- src/main/java/io/github/mzmine/datamodel/features/types/GNPSSpectralLibMatchSummaryType.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2006-2020 The MZmine Development Team
- *
- * This file is part of MZmine.
- *
- * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
- *
- * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
- * USA
- */
-
-package io.github.mzmine.datamodel.features.types;
-
-import io.github.mzmine.datamodel.features.types.modifiers.AnnotationType;
-import io.github.mzmine.datamodel.features.types.modifiers.EditableColumnType;
-import io.github.mzmine.datamodel.features.types.numbers.abstr.ListDataType;
-import io.github.mzmine.modules.dataprocessing.id_gnpsresultsimport.GNPSLibraryMatch;
-import io.github.mzmine.modules.dataprocessing.id_gnpsresultsimport.GNPSResultsImportModule;
-import org.jetbrains.annotations.NotNull;
-
-/**
- * A list of spectral library matches imported from GNPS results (see {@link
- * GNPSResultsImportModule}
- */
-public class GNPSSpectralLibMatchSummaryType extends ListDataType<GNPSLibraryMatch>
-    implements AnnotationType, EditableColumnType {
-
-  @NotNull
-  @Override
-  public String getHeaderString() {
-    return "GNPS";
-  }
-
-}
