+++ src/main/java/io/github/mzmine/datamodel/features/types/numbers/FormulaMassType.java
@@ -18,7 +18,7 @@
 
 package io.github.mzmine.datamodel.features.types.numbers;
 
-import io.github.mzmine.datamodel.features.types.FormulaType;
+import io.github.mzmine.datamodel.features.types.annotations.FormulaType;
 import io.github.mzmine.datamodel.features.types.numbers.abstr.DoubleType;
 import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.util.FormulaUtils;
