+++ src/main/java/io/github/mzmine/datamodel/features/types/numbers/MzPpmDifferenceType.java
@@ -18,7 +18,7 @@
 
 package io.github.mzmine.datamodel.features.types.numbers;
 
-import io.github.mzmine.datamodel.features.types.FormulaAnnotationType;
+import io.github.mzmine.datamodel.features.types.annotations.FormulaAnnotationType;
 import io.github.mzmine.datamodel.features.types.numbers.abstr.FloatType;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
