+++ src/main/java/io/github/mzmine/modules/dataprocessing/id_gnpsresultsimport/GNPSResultsImportTask.java
@@ -25,8 +25,8 @@ import io.github.mzmine.datamodel.features.SimpleFeatureListAppliedMethod;
 import io.github.mzmine.datamodel.features.correlation.R2RMS2CosineSimilarityGNPS;
 import io.github.mzmine.datamodel.features.correlation.R2RMap;
 import io.github.mzmine.datamodel.features.correlation.RowsRelationship.Type;
-import io.github.mzmine.datamodel.features.types.GNPSSpectralLibMatchSummaryType;
-import io.github.mzmine.datamodel.features.types.GNPSSpectralLibraryMatchType;
+import io.github.mzmine.datamodel.features.types.annotations.GNPSSpectralLibMatchSummaryType;
+import io.github.mzmine.datamodel.features.types.annotations.GNPSSpectralLibraryMatchType;
 import io.github.mzmine.modules.dataprocessing.id_gnpsresultsimport.GNPSLibraryMatch.ATT;
 import io.github.mzmine.parameters.ParameterSet;
 import io.github.mzmine.taskcontrol.AbstractTask;
